# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Сезоны — API дашборда",
    "description": "Аналитика продаж ресторана из iiko.\n\nКонтракты: бэкенд отдаёт факты, представление — зона фронтенда. docs/frontend-handoff.md",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/sales/snapshot": {
      "get": {
        "tags": [
          "Продажи"
        ],
        "summary": "Снимок продаж: период + позиции (1 RTT)",
        "operationId": "getSalesSnapshot",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SalesPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sales/positions": {
      "get": {
        "tags": [
          "Продажи"
        ],
        "summary": "Позиции продаж за период",
        "operationId": "getSalesPositions",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SalesPosition"
                  },
                  "title": "Response Getsalespositions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/bounds": {
      "get": {
        "tags": [
          "Склад"
        ],
        "summary": "Края истории слепков склада",
        "operationId": "getStockBounds",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Getstockbounds"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/stock/snapshot": {
      "get": {
        "tags": [
          "Склад"
        ],
        "summary": "Снимок склада: totals + positions + dynamics (1 RTT)",
        "operationId": "getStockSnapshot",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Начало dynamics",
              "title": "Date From"
            },
            "description": "Начало dynamics"
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Конец dynamics",
              "title": "Date To"
            },
            "description": "Конец dynamics"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/totals": {
      "get": {
        "tags": [
          "Склад"
        ],
        "summary": "Get Stock Totals",
        "operationId": "getStockTotals",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreValue"
                  },
                  "title": "Response Getstocktotals"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/positions": {
      "get": {
        "tags": [
          "Склад"
        ],
        "summary": "Get Stock Positions",
        "operationId": "getStockPositions",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StockPosition"
                  },
                  "title": "Response Getstockpositions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/negative": {
      "get": {
        "tags": [
          "Склад"
        ],
        "summary": "Get Stock Negative",
        "operationId": "getStockNegative",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegativeStock"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/dynamics": {
      "get": {
        "tags": [
          "Склад"
        ],
        "summary": "Get Stock Dynamics",
        "operationId": "getStockDynamics",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DynamicsPoint"
                  },
                  "title": "Response Getstockdynamics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DynamicsPoint": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "description": "День слепка"
          },
          "byStore": {
            "items": {
              "$ref": "#/components/schemas/StoreValue"
            },
            "type": "array",
            "title": "Bystore",
            "description": "Всегда все три юнита k/b/w (нулевые включены); total фронт суммирует сам"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "date",
          "byStore"
        ],
        "title": "DynamicsPoint",
        "description": "Одна точка ГРАФИКА «стоимость склада во времени» (блок dynamics).\n\nЗачем: остатки — замороженные деньги, ценность в ТРЕНДЕ, не в разовом\nчисле. Линия вверх = закупаем быстрее, чем расходуем (деньги мертвеют\nна полках, риск порчи — виден ДО инвентаризации); ступенька вверх =\nкрупная закупка (можно открыть слепок того дня). Разрез по юнитам\nв каждой точке — чтобы фронт рисовал и общую линию (сумма byStore),\nи линии по складам из одних данных. Гранулярность неделя/месяц —\nпрореживание точек на фронте, не отдельные данные."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NegativeStock": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count",
            "description": "Позиций с отрицательным остатком"
          },
          "valueAbs": {
            "type": "number",
            "title": "Valueabs",
            "description": "Суммарная «дыра» по модулю (сколько списано сверх оприходованного в деньгах). Округлено до 2 знаков"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "count",
          "valueAbs"
        ],
        "title": "NegativeStock",
        "description": "Готовое предупреждение о минусовых остатках.\n\nПравило «минус не входит в totals» применяет бэкенд. Полный список\nминусовых строк — фильтр positions по qty < 0 (они там есть).\nНоль везде — честный ноль: учёт чистый."
      },
...
      "SalesPage": {
        "properties": {
          "period": {
            "$ref": "#/components/schemas/src__schemas__sales__Period"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/SalesPosition"
            },
            "type": "array",
            "title": "Positions",
            "description": "Агрегат по блюдам за период; позиции с нулевой оплатой (проработки, включённые в банкет) исключены"
          }
        },
        "type": "object",
        "required": [
          "period",
          "positions"
        ],
        "title": "SalesPage"
      },
      "SalesPosition": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Название блюда (как в iiko, пробелы подстрижены)"
          },
          "sub": {
            "type": "string",
            "title": "Sub",
            "description": "Подкатегория для второго уровня детализации (категория блюда из iiko)"
          },
          "cat": {
            "type": "string",
            "title": "Cat",
            "description": "Юнит: k кухня, b бар, w вино, o «вне подразделений» (в донат не входит, в таблице и общей выручке — участвует)"
          },
          "qty": {
            "type": "number",
            "title": "Qty",
            "description": "Продано порций за период; дробное у весовых блюд (4.5 кг)"
          },
          "revenue": {
            "type": "number",
            "title": "Revenue",
            "description": "Выручка позиции за период — СУММА фактических оплат (не произведение qty*price: точность не теряется на округлении)"
          },
          "listValue": {
            "type": "number",
            "title": "Listvalue",
            "description": "Прейскурантная выручка (Σ цен меню без скидок): цена меню = listValue/qty (круглая, если цена в периоде не менялась); скидки позиции = listValue − revenue"
          },
          "cost": {
            "type": "number",
            "title": "Cost",
            "description": "Суммарная себестоимость проданного за период; 0 = техкарты нет (fc и gp позиции показывать «—», не 0%)"
          },
          "price": {
            "type": "number",
            "title": "Price",
            "description": "Средняя фактическая цена порции (со скидками, = выручка/qty)"
          },
          "unitCost": {
            "type": "number",
            "title": "Unitcost",
            "description": "Средняя себестоимость порции"
          }
        },
        "type": "object",
        "required": [
          "name",
          "sub",
          "cat",
          "qty",
          "revenue",
          "listValue",
          "cost",
          "price",
          "unitCost"
        ],
        "title": "SalesPosition",
        "description": "Строка таблицы позиций. Производные считает фронт:\nprice = rev/qty, unitCost= cost/qty, gp = rev-cost, fc = cost/rev."
      },
...
      "StockDataBounds": {
        "properties": {
          "earliest": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earliest",
            "description": "Самый ранний слепок (глубина бэкфила) = min(availableDates); null — слепков нет. Для нижней границы календаря"
          },
          "latest": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest",
            "description": "Самый свежий слепок (последний закрытый день) = max(availableDates); null — слепков нет. Дефолт asOf и верхняя граница календаря"
          },
          "availableDates": {
            "items": {
              "type": "string",
              "format": "date"
            },
            "type": "array",
            "title": "Availabledates",
            "description": "Точный список дней, за которые слепок есть — day-picker гасит даты вне списка. Кодирует и ретеншн (воскресенья глубже года), и дыры (нерабочие дни, сбои синка)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "earliest",
          "latest",
          "availableDates"
        ],
        "title": "StockDataBounds",
        "description": "Контракт day-picker'а: какие дни можно выбрать.\n\nearliest/latest задают границы календаря, availableDates — точный\nнабор пикабельных дней внутри них. Day-picker гасит всё, чего нет\nв availableDates, поэтому запрос ?date всегда попадает в имеющийся\nслепок (404 при корректном пикере не случается).\n\nРетеншн виден прямо в availableDates: свежая история — все дни,\nстарше скользящего года — только воскресенья; плюс возможны дыры\n(заведение не работало, синк не прошёл)."
      },
...
      "StockPosition": {
        "properties": {
          "productId": {
            "type": "string",
            "title": "Productid",
            "description": "UUID продукта в iiko — стабильная идентичность позиции: имя может меняться, id — никогда. Связь позиции между слепками разных дней — только по нему"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Продукт — имя на момент синка. При переименовании в iiko новые слепки получают новое имя, старые хранят старое (история как была); идентичность — productId"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "Категория: ближайшая папка номенклатуры iiko (тоже слепок на момент синка)"
          },
          "store": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w"
            ],
            "title": "Store",
            "description": "Юнит склада"
          },
          "qty": {
            "type": "number",
            "title": "Qty",
            "description": "Остаток; дробное у весовых (4.5 кг); может быть ОТРИЦАТЕЛЬНЫМ — грязный складской учёт (см. negativeStock). Округлено до 3 знаков (точность весовых)"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Единица измерения: кг / л / шт / уп / …"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Стоимость остатка; цена за единицу = value / qty (фронт). Округлено до 2 знаков — чтобы производная цена за единицу не теряла точность"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "productId",
          "name",
          "category",
          "store",
          "qty",
          "unit",
          "value"
        ],
        "title": "StockPosition",
        "description": "Позиция слепка дня — как есть в учёте, ВКЛЮЧАЯ отрицательные."
      },
      "StockSnapshot": {
        "properties": {
          "asOf": {
            "type": "string",
            "format": "date",
            "title": "Asof",
            "description": "День показываемого слепка (конец этого закрытого дня); выбирается ?date, дефолт — latest. ?date обязан быть из dataBounds.availableDates, иначе 404"
          },
          "dataBounds": {
            "$ref": "#/components/schemas/StockDataBounds",
            "description": "Контракт day-picker'а: границы календаря + точный список пикабельных дней (availableDates)"
          },
          "totals": {
            "items": {
              "$ref": "#/components/schemas/StoreValue"
            },
            "type": "array",
            "title": "Totals",
            "description": "Всегда три элемента k/b/w (нулевые включены); только положительные остатки"
          },
          "positions": {
            "items": {
              "$ref": "#/components/schemas/StockPosition"
            },
            "type": "array",
            "title": "Positions",
            "description": "ПОЛНЫЙ состав слепка, включая минусовые строки (~300–700 позиций), по убыванию value; сортировка/поиск — фронт. ВНИМАНИЕ: сумма positions ≠ totals — totals считает только положительные строки"
          },
          "negativeStock": {
            "$ref": "#/components/schemas/NegativeStock"
          },
          "dynamics": {
            "items": {
              "$ref": "#/components/schemas/DynamicsPoint"
            },
            "type": "array",
            "title": "Dynamics",
            "description": "ДАННЫЕ ГРАФИКА динамики стоимости склада — единственный источник линии (снапшот-блоки её не кормят). Точки по дням за ?from/?to (дефолт — 30 дней до asOf); режим недели/месяца — прореживание на фронте"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "asOf",
          "dataBounds",
          "totals",
          "positions",
          "negativeStock",
          "dynamics"
        ],
        "title": "StockSnapshot"
      },
...
      "StoreValue": {
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "k",
              "b",
              "w"
            ],
            "title": "Key",
            "description": "Юнит склада: k кухня, b бар, w вино. Источник — маппинг физических складов при онбординге (по id); стандарт требует склады зеркально продажам (паспорт §9). В деньгах со своим юнитом продаж не сходится: сырьё одного блюда может хранится на разных складах"
          },
          "value": {
            "type": "number",
            "title": "Value",
            "description": "Стоимость ПОЛОЖИТЕЛЬНЫХ остатков; минусовые строки сюда не входят (см. negativeStock). Округлено до целой единицы валюты (агрегатные деньги)"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "key",
          "value"
        ],
        "title": "StoreValue",
        "description": "Стоимость остатков одного юнита складов."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "src__schemas__sales__Period": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label",
            "description": "LEGACY: готовая подпись периода (\"02.06 — 01.07.2026\"); умирает в такте 5 — фронт будет форматировать сам из dateFrom/dateTo"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "description": "LEGACY: подстрока-пояснение, умирает вместе с label"
          },
          "dateFrom": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datefrom",
            "description": "Начало ФАКТИЧЕСКИ покрытого периода — после дефолта и усечения краями данных (кейс: запросили по 31-е, закрыто по 16-е). null — данных нет вовсе"
          },
          "dateTo": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateto",
            "description": "Конец фактически покрытого периода (последний закрытый день при усечении); null — данных нет"
          }
        },
        "type": "object",
        "required": [
          "label",
          "note"
        ],
        "title": "Period"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}